requirements.txt
# RiskMapper Agent System - Dependencies
# Python 3.11+ required

# Core
pydantic>=2.0
python-dotenv>=1.0.0
requests>=2.31.0

# Vector Store
chromadb>=0.4.0

# Excel parsing
openpyxl>=3.1.0

# Testing
pytest>=7.0
hypothesis>=6.0
pytest-mock>=3.0

# Logging (stdlib - no install needed)
# json (stdlib - no install needed)
